requirements.txt
flask
flask-socketio
flask-login
flask-sqlalchemy
flask_jwt_extended
eventlet
werkzeug
psycopg2-binary
python-dotenv